+++ CHANGES.md
@@ -10,6 +10,7 @@ New:
 Fixed:
 
 - Optimized CPU usage (#4369, #4370)
+- Fixed empty initial HLS segment (#4401)
 - Fixed support for `duration` metadata in image decoder (#4397)
 - Fixed cue-out bug in cue file parsing (#4381)
 - Bring back parse error location. (#4362)